Team_Magnolia/coursework_two/pyproject.toml
[tool.poetry]
name = "coursework-two"
version = "0.1.0"
description = "ESG Report Analyzer for CSR Indicators"
authors = ["TeamMagnolia"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
pymupdf = "^1.25.5"
docling = "^2.30.0"
openai = "^1.25.0"
pymongo = "^4.12.0"

# —— 新增 —— 可视化 & 重试 & 指标 & PostgreSQL 驱动
flask = "^2.2.5"
prometheus-client = "^0.16.0"
tenacity = "^8.2.2"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
bandit = "^1.7.5"

# —— 新增 —— 安全扫描 & MongoDB Mock
safety = "^2.3.1"
mongomock = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=modules --cov-report=term-missing --cov-report=html"
